pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-email-triage"
version = "1.0.0"
description = "Inbox management RL environment with three tasks of increasing difficulty."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "OpenEnv Team"}
]
keywords = ["openenv", "reinforcement-learning", "email-triage", "environment"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pydantic==2.7.1",
    "fastapi==0.111.0",
    "uvicorn==0.29.0",
    "hypothesis==6.100.0",
    "openai==1.25.0",
    "pyyaml==6.0.1",
    "pytest==8.2.0",
    "httpx==0.27.0",
    "python-multipart==0.0.9",
    "aiofiles==23.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "hypothesis>=6.100.0",
    "httpx>=0.27.0",
]

[project.urls]
Homepage = "https://huggingface.co/spaces/Superpredater231/openenv-email-triage"
Repository = "https://github.com/Superpredater/meta-project-2026"

[tool.setuptools]
packages = ["openenv_email_triage"]

[tool.setuptools.package-data]
openenv_email_triage = ["fixtures/*.json"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"

[tool.coverage.run]
source = ["openenv_email_triage"]
omit = ["tests/*", "*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
